tests for schedule create
